+++ src/main/java/com/aetherteam/aether/accessories/client/gui/ToggleButton.java
@@ -1,8 +1,8 @@
 package com.aetherteam.aether.accessories.client.gui;
 
 import com.aetherteam.aether.accessories.api.menu.AccessoriesBasedSlot;
-import com.aetherteam.aether.accessories.networking.AccessoriesNetworking;
 import com.aetherteam.aether.accessories.networking.server.ToggleAccessoryRenderPacket;
+import com.aetherteam.aether.network.AetherPacketSender;
 import net.minecraft.client.gui.GuiGraphicsExtractor;
 import net.minecraft.client.gui.components.Button;
 import net.minecraft.network.chat.Component;
@@ -14,7 +14,7 @@ public class ToggleButton extends Button {
         super(x, y, 12, 12, Component.empty(), (button) -> {
             boolean shouldRender = !slot.shouldRender();
             slot.setRender(shouldRender);
-            AccessoriesNetworking.sendToServer(new ToggleAccessoryRenderPacket(slot.slotName(), slot.slotIndex(), shouldRender));
+            AetherPacketSender.sendToServer(new ToggleAccessoryRenderPacket(slot.slotName(), slot.slotIndex(), shouldRender));
         }, DEFAULT_NARRATION);
         this.slot = slot;
     }